[#105] Convert button icons file into class
